requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130
torch==2.10.0+cu130
torchao==0.16.0
torchvision==0.25.0+cu130
transformers==5.1.0
PyYAML==6.0.3
tqdm==4.67.*
pandas==3.0.0
pyarrow==23.0.0
rich==14.3.2
ninja==1.13.0